Code folder and files restructuring
segregated pages, components and styles
