+++ src/component/Footer/Footer.jsx
@@ -1,6 +1,5 @@
 import React from "react";
-// import { Link } from "react-router-dom";
-import "./Footer.css";
+import "../../styles/Footer.css";
 
 export default function Footer() {
   return (
