+++ src/component/Header/Header.jsx
@@ -1,6 +1,6 @@
 import React, { useState } from "react";
 import { Link, NavLink } from "react-router-dom";
-import "./Header.css";
+import "../../styles/Header.css";
 
 export default function Header() {
   const [menuOpen, setMenuOpen] = useState(false);
