--- src/component/Home/HomePage.jsx
@@ -1,18 +1,0 @@
-import React from "react";
-import Home from "./Home";
-import Programs from "./Programs";
-import TopCities from "./TopCities";
-import HomeBlog from "./HomeBlog";
-
-const HomePage = () => {
-  return (
-    <>
-      <Home />
-      <Programs />
-      <TopCities />
-      <HomeBlog />
-    </>
-  );
-};
-
-export default HomePage;
