+++ src/styles/Colleges/Colleges.css
