+++ src/styles/Colleges/Ranking.css
